Src/kg-service/requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
python-dotenv==1.2.1
langchain==1.2.10
langchain-community==0.4.1
langchain-neo4j==0.8.0
neo4j==6.1.0
networkx==3.6.1
pydantic>=2.0.0
pandas>=2.0.0
pydantic-settings>=2.0.0
pytest